third_party/blink/web_tests/wpt_internal/css/css-contain/container-queries/scroll-state-snapped-container-type-change-expected.txt

This is a testharness.js-based test.
[FAIL] Check scroll-state(snapped) not matching when container-type is removed
  assert_equals: container-type re-added expected "yes" but got "no"
Harness: the test ran to completion.
